Change HikariPooledConnectionProvider constructors, add SlidingExpirationMap javadocs
